backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.0
langchain-core==0.3.0
langchain-openai==0.2.0
supabase==2.8.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
redis==5.0.0
